Move ExcelFileGenerator to separate file
Renamed the FileGenerator to ExcelFileGenerator and moved it to separate file.
